+++ BASE/src/ServerTelemetryChannel/Implementation/PartialSuccessTransmissionPolicy.cs
@@ -101,21 +101,24 @@
             if (args.Response != null && args.Response.StatusCode == ResponseStatusCodes.PartialSuccess)
             {
                 int statusCode;
+                Transmission transmission;
                 string newTransmissions = ParsePartialSuccessResponse(args.Transmission, args, out statusCode);
 
                 if (!string.IsNullOrEmpty(newTransmissions))
                 {
-                    this.DelayFutureProcessing(args.Response, statusCode);
-                    
-                    byte[] data = JsonSerializer.ConvertToByteArray(newTransmissions);
-                    Transmission newTransmission = new Transmission(
-                        args.Transmission.EndpointAddress,
-                        data,
-                        args.Transmission.ContentType,
-                        args.Transmission.ContentEncoding,
-                        args.Transmission.Timeout);
+                    if (args.Transmission.IsFlushAsyncInProgress)
+                    {
+                        // Move newTransmission to storage on IAsyncFlushable.FlushAsync
+                        transmission = this.SerializeNewTransmission(args, newTransmissions);
+                        this.DelayFutureProcessing(args.Response, statusCode);
+                    }
+                    else
+                    {
+                        this.DelayFutureProcessing(args.Response, statusCode);
+                        transmission = this.SerializeNewTransmission(args, newTransmissions);
+                    }
 
-                    this.Transmitter.Enqueue(newTransmission);
+                    this.Transmitter.Enqueue(transmission);
                 }
                 else
                 {
@@ -123,6 +126,19 @@
                     this.backoffLogicManager.ResetConsecutiveErrors();
                 }
             }
+        }
+
+        private Transmission SerializeNewTransmission(TransmissionProcessedEventArgs args, string newTransmissions)
+        {
+            byte[] data = JsonSerializer.ConvertToByteArray(newTransmissions);
+            Transmission transmission = new Transmission(
+                args.Transmission.EndpointAddress,
+                data,
+                args.Transmission.ContentType,
+                args.Transmission.ContentEncoding,
+                args.Transmission.Timeout);
+
+            return transmission;
         }
 
         private void DelayFutureProcessing(HttpWebResponseWrapper response, int statusCode)