+++ BASE/src/ServerTelemetryChannel/Implementation/TelemetrySerializer.cs
@@ -3,6 +3,8 @@
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using System.Threading;
+    using System.Threading.Tasks;
     using Microsoft.ApplicationInsights.Channel;
 
     internal class TelemetrySerializer
@@ -48,9 +50,35 @@
                 throw new Exception("TelemetrySerializer.EndpointAddress was not set.");
             }
 
-            var transmission = new Transmission(this.EndpointAddress, items) 
-                                        { TransmissionStatusEvent = this.TransmissionStatusEvent };
+            var transmission = new Transmission(this.EndpointAddress, items)
+                                    { TransmissionStatusEvent = this.TransmissionStatusEvent };
             this.Transmitter.Enqueue(transmission);
+        }
+
+        public virtual Task<bool> SerializeAsync(ICollection<ITelemetry> items, CancellationToken cancellationToken)
+        {
+            if (this.EndpointAddress == null)
+            {
+                throw new Exception("TelemetrySerializer.EndpointAddress was not set.");
+            }
+
+            if (items == null || !items.Any())
+            {
+                return this.Transmitter.MoveTransmissionsAndWaitForSender(cancellationToken);
+            }
+
+            var transmission = new Transmission() { TransmissionStatusEvent = this.TransmissionStatusEvent };
+            // serialize transmission and enqueue asynchronously
+            return Task.Run(() => this.SerializeTransmissionAndEnqueue(transmission, items, cancellationToken), cancellationToken);
+        }
+
+        /// <summary>
+        /// Serializes transmission and enqueue.
+        /// </summary>
+        private Task<bool> SerializeTransmissionAndEnqueue(Transmission transmission, ICollection<ITelemetry> items, CancellationToken cancellationToken)
+        {
+            transmission.Serialize(this.EndpointAddress, items);
+            return this.Transmitter.FlushAsync(transmission, cancellationToken);
         }
     }
 }